backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "fastapi-mail>=1.6.1",
    "httpx>=0.28.1",
    "langchain>=1.2.7",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.0",
    "logger>=1.4",
    "loguru>=0.7.3",
    "mistune>=3.2.0",
    "numpy>=2.4.1",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "redis>=7.2.1",
    "scikit-learn>=1.8.0",
    "spacy>=3.8.11",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "yfinance>=1.2.0",
]
